第3章需求

+++ open-core-facade/src/main/java/com/sm/open/core/facade/model/param/pf/biz/kb/assess/FaqEvaCaseItemParam.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.ToString;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 知识库_病例评估_评估项
@@ -43,7 +44,7 @@ public class FaqEvaCaseItemParam implements Serializable {
     /**
      * 评估项分值
      */
-    private Integer scoreEva;
+    private BigDecimal scoreEva;
 
     /**
      * 评估阶段
